功能测试中, 正在进行数据库格式修改

+++ common/src/constants/redis_keys/user.rs
@@ -1,5 +1,5 @@
 #[inline]
-pub fn user_access_token(user_id: u32) -> String {
+pub fn user_access_token(user_id: i64) -> String {
     format!("auth:user:accessToken:{}", user_id)
 }
 #[inline]
@@ -11,6 +11,6 @@ pub fn inviter_code(code: &str) -> String {
     format!("auth:inviter:code:{}", code)
 }
 #[inline]
-pub fn user_info_cache(user_id: u32) -> String {
+pub fn user_info_cache(user_id: i64) -> String {
     format!("auth:user:info:cache:{}", user_id)
 }
