+++ entities/src/collection.rs
@@ -5,12 +5,12 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "photo_collection")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub user_id: i32,
+    pub id: i64,
+    pub user_id: i64,
     pub name: String,
     pub description: Option<String>,
-    pub photo_count: i32,
-    pub cover_image_id: Option<i32>,
+    pub photo_count: i64,
+    pub cover_image_id: Option<i64>,
     pub is_favorite: bool,
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
